CyberKidsFinal/kidproc2/kid_protocol_v2/requirements.txt
httplib2==0.18.1
requests==2.23.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
json2html==1.3.0
simplejson==3.17.0
PyYAML==3.13
urllib3==1.25.9
zipp==1.0.0
packaging==20.9
python-dateutil==2.8.1
